requirements.txt
# Core dependencies
aiogram==3.15.0
aiohttp==3.11.11
python-dotenv==1.0.1

# LangChain and AI
langchain==0.3.13
langchain-openai==0.2.14
openai==1.58.1

# Database
supabase==2.11.2

# Memory and context
zep-cloud==1.1.3

# Audio processing
pydub==0.25.1

# Utilities
pydantic==2.10.5
pydantic-settings==2.7.0
aiofiles==24.1.0
python-multipart==0.0.21

# Development tools
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1
mypy==1.14.1

# Logging and monitoring
loguru==0.7.3
prometheus-client==0.21.1